requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Django==2.1.2
django-filter==2.0.0
et-xmlfile==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
jdcal==1.4
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.0
mysqlclient==1.3.13
oauthlib==2.0.6
pycparser==2.19
pytz==2018.7
six==1.11.0
